Split extensions into multiple modules
