+++ lib/postgrex/utils.ex
@@ -1,6 +1,31 @@
 defmodule Postgrex.Utils do
   @moduledoc false
 
+  @extensions [
+    Postgrex.Extensions.Array,
+    Postgrex.Extensions.Bool,
+    Postgrex.Extensions.Date,
+    Postgrex.Extensions.Float4,
+    Postgrex.Extensions.Float8,
+    Postgrex.Extensions.HStore,
+    Postgrex.Extensions.Int2,
+    Postgrex.Extensions.Int4,
+    Postgrex.Extensions.Int8,
+    Postgrex.Extensions.Interval,
+    Postgrex.Extensions.MACADDR,
+    Postgrex.Extensions.Network,
+    Postgrex.Extensions.Numeric,
+    Postgrex.Extensions.OID,
+    Postgrex.Extensions.Range,
+    Postgrex.Extensions.Raw,
+    Postgrex.Extensions.Record,
+    Postgrex.Extensions.TID,
+    Postgrex.Extensions.Time,
+    Postgrex.Extensions.Timestamp,
+    Postgrex.Extensions.Interval,
+    Postgrex.Extensions.UUID,
+    Postgrex.Extensions.Void]
+
   @doc """
   Converts pg major.minor.patch (http://www.postgresql.org/support/versioning) version to an integer
   """
@@ -29,4 +54,28 @@ defmodule Postgrex.Utils do
     |> Keyword.put_new(:port, System.get_env("PGPORT"))
     |> Enum.reject(fn {_k, v} -> is_nil(v) end)
   end
+
+  @doc """
+  List all default extensions.
+  """
+  def default_extensions() do
+    unquote(Enum.map(@extensions, &{&1, nil}))
+  end
+
+  @doc """
+  Return encode error message.
+  """
+  def encode_msg(%Postgrex.TypeInfo{type: type}, observed, expected) do
+    "Postgrex expected #{to_desc(expected)} that can be encoded/cast to " <>
+    "type #{inspect type}, got #{inspect observed}. Please make sure the " <>
+    "value you are passing matches the definition in your table or in your " <>
+    "query or convert the value accordingly."
+  end
+
+  ## Helpers
+
+  defp to_desc(struct) when is_atom(struct), do: "%#{inspect struct}{}"
+  defp to_desc(%Range{} = range), do: "an integer in #{inspect range}"
+  defp to_desc({a, b}), do: to_desc(a) <> " or " <> to_desc(b)
+  defp to_desc(desc) when is_binary(desc), do: desc
 end
